[FIX] SERVICE SUMMARYDTO.MONTHLYSUMMARY 변경완 REPOSITORYSUPPORT쪽 겹치는 거 합치기 완료 달성률 int -> double 변경환

+++ src/main/java/com/kurung/exercise/dto/SummaryDTO.java
@@ -22,7 +22,7 @@ public class SummaryDTO {
   protected int totalDuration;
   protected int totalKcal;
   protected int routineCount;
-  protected int goalAchievementRate;
+  protected double goalAchievementRate;
   protected List<SummaryDTO.ExerciseLogDTO> exerciseList;
 
   @Getter
